Add files via upload
The primary objective of this analysis is to build a Deep Learning model for classifying images in the CIFAR-10 dataset. The analysis will explore the performance of three different models to determine which architecture is most effective for image classification:
Sequential CNN – Basic CNN model for baseline performance.
Deeper CNN with Dropout – A more complex CNN with added regularization to reduce overfitting.
Pre-trained ResNet – Transfer learning using a pre-trained ResNet model for better feature extraction.